added features for chargers

+++ timeslots/views.py
@@ -81,6 +81,9 @@ def slotstatus(request, slot_id, station_id, date):
         if slot.progress < 4:
             slot.progress = F('progress') + 1
             slot.save()
+        else:
+            slot.progress = 0 
+            slot.save()
         return HttpResponseRedirect('/timeslots/station/%s/date/%s/slots/' % (station_id, date))
     else:
         log_task(request, "User %s tried to change the status of slot %s but is not allowed to." % (request.user, slot))
@@ -290,6 +293,10 @@ def slot(request, date, block_id, timeslot, line):
                     defaults={'company': request.user.userprofile})
 
     # check conditions
+    if request.user.userprofile.is_charger:
+        log_task(request, "Charger %s tried to access slot %s." % (request.user, slot))
+        messages.error(request, _('You are not allowed to change this slot!'))
+        return HttpResponseRedirect('/timeslots/station/%s/date/%s/jobtable/' % (block.dock.station.id, date))
     if not slot.block.dock.station.opened_on_weekend and not request.user.userprofile.is_master and datetime.strptime(date, "%Y-%m-%d").date().weekday() > 4:
         if created:
             slot.delete()
